Move SharpPcap.Npcap.OpenFlags to SharpPcap.OpenFlags
