+++ Examples/Example3.BasicCap/Program.cs
@@ -56,7 +56,7 @@ namespace Example3
             if (device is NpcapDevice)
             {
                 var nPcap = device as NpcapDevice;
-                nPcap.Open(SharpPcap.Npcap.OpenFlags.DataTransferUdp | SharpPcap.Npcap.OpenFlags.NoCaptureLocal, readTimeoutMilliseconds);
+                nPcap.Open(OpenFlags.DataTransferUdp | OpenFlags.NoCaptureLocal, readTimeoutMilliseconds);
             }
             else if (device is LibPcapLiveDevice)
             {
